TimelockController refactor & cleanup
