+++ test/access/TimelockController.test.js
@@ -7,7 +7,7 @@ const { expect } = require('chai');
 const TimelockController = contract.fromArtifact('TimelockController');
 const CallReceiverMock = contract.fromArtifact('CallReceiverMock');
 const Implementation2 = contract.fromArtifact('Implementation2');
-const MINDELAY = time.duration.days(1);
+const MINDELAY = web3.utils.toBN(web3.utils.randomHex(4));
 
 function genOperation (target, value, data, predecessor, salt) {
   const id = web3.utils.keccak256(web3.eth.abi.encodeParameters([
@@ -86,13 +86,13 @@ describe('TimelockController', function () {
             MINDELAY,
             { from: proposer }
           );
-          expectEvent(receipt, 'Scheduled', { id: this.operation.id });
           expectEvent(receipt, 'CallScheduled', {
             id: this.operation.id,
             index: web3.utils.toBN(0),
             target: this.operation.target,
             value: web3.utils.toBN(this.operation.value),
             data: this.operation.data,
+            predecessor: this.operation.predecessor,
           });
         });
 
@@ -106,13 +106,13 @@ describe('TimelockController', function () {
             MINDELAY,
             { from: proposer }
           );
-          expectEvent(receipt, 'Scheduled', { id: this.operation.id });
           expectEvent(receipt, 'CallScheduled', {
             id: this.operation.id,
             index: web3.utils.toBN(0),
             target: this.operation.target,
             value: web3.utils.toBN(this.operation.value),
             data: this.operation.data,
+            predecessor: this.operation.predecessor,
           });
 
           const block = await web3.eth.getBlock(receipt.receipt.blockHash);
@@ -131,13 +131,13 @@ describe('TimelockController', function () {
             MINDELAY,
             { from: proposer }
           );
-          expectEvent(receipt, 'Scheduled', { id: this.operation.id });
           expectEvent(receipt, 'CallScheduled', {
             id: this.operation.id,
             index: web3.utils.toBN(0),
             target: this.operation.target,
             value: web3.utils.toBN(this.operation.value),
             data: this.operation.data,
+            predecessor: this.operation.predecessor,
           });
 
           await expectRevert(
@@ -150,7 +150,7 @@ describe('TimelockController', function () {
               MINDELAY,
               { from: proposer }
             ),
-            'Timelock: operation already scheduled'
+            'TimelockController: operation already scheduled'
           );
         });
 
@@ -192,14 +192,14 @@ describe('TimelockController', function () {
                 this.operation.salt,
                 { from: executer }
               ),
-              'Timelock: operation is not ready'
+              'TimelockController: operation is not ready'
             );
           });
         });
 
         describe('with scheduled operation', function () {
           beforeEach(async function () {
-            ({ logs: this.logs } = await this.tlctrl.schedule(
+            ({ receipt: this.receipt, logs: this.logs } = await this.tlctrl.schedule(
               this.operation.target,
               this.operation.value,
               this.operation.data,
@@ -221,7 +221,7 @@ describe('TimelockController', function () {
                   this.operation.salt,
                   { from: executer }
                 ),
-                'Timelock: operation is not ready'
+                'TimelockController: operation is not ready'
               );
             });
           });
@@ -242,7 +242,7 @@ describe('TimelockController', function () {
                   this.operation.salt,
                   { from: executer }
                 ),
-                'Timelock: operation is not ready'
+                'TimelockController: operation is not ready'
               );
             });
           });
@@ -262,7 +262,6 @@ describe('TimelockController', function () {
                 this.operation.salt,
                 { from: executer }
               );
-              expectEvent(receipt, 'Executed', { id: this.operation.id });
               expectEvent(receipt, 'CallExecuted', {
                 id: this.operation.id,
                 index: web3.utils.toBN(0),
@@ -312,7 +311,6 @@ describe('TimelockController', function () {
             MINDELAY,
             { from: proposer }
           );
-          expectEvent(receipt, 'Scheduled', { id: this.operation.id });
           for (const i in this.operation.targets) {
             expectEvent(receipt, 'CallScheduled', {
               id: this.operation.id,
@@ -320,6 +318,7 @@ describe('TimelockController', function () {
               target: this.operation.targets[i],
               value: web3.utils.toBN(this.operation.values[i]),
               data: this.operation.datas[i],
+              predecessor: this.operation.predecessor,
             });
           }
         });
@@ -334,7 +333,6 @@ describe('TimelockController', function () {
             MINDELAY,
             { from: proposer }
           );
-          expectEvent(receipt, 'Scheduled', { id: this.operation.id });
           for (const i in this.operation.targets) {
             expectEvent(receipt, 'CallScheduled', {
               id: this.operation.id,
@@ -342,6 +340,7 @@ describe('TimelockController', function () {
               target: this.operation.targets[i],
               value: web3.utils.toBN(this.operation.values[i]),
               data: this.operation.datas[i],
+              predecessor: this.operation.predecessor,
             });
           }
 
@@ -361,7 +360,6 @@ describe('TimelockController', function () {
             MINDELAY,
             { from: proposer }
           );
-          expectEvent(receipt, 'Scheduled', { id: this.operation.id });
           for (const i in this.operation.targets) {
             expectEvent(receipt, 'CallScheduled', {
               id: this.operation.id,
@@ -369,6 +367,7 @@ describe('TimelockController', function () {
               target: this.operation.targets[i],
               value: web3.utils.toBN(this.operation.values[i]),
               data: this.operation.datas[i],
+              predecessor: this.operation.predecessor,
             });
           }
 
@@ -382,7 +381,7 @@ describe('TimelockController', function () {
               MINDELAY,
               { from: proposer }
             ),
-            'Timelock: operation already scheduled'
+            'TimelockController: operation already scheduled'
           );
         });
 
@@ -424,14 +423,14 @@ describe('TimelockController', function () {
                 this.operation.salt,
                 { from: executer }
               ),
-              'Timelock: operation is not ready'
+              'TimelockController: operation is not ready'
             );
           });
         });
 
         describe('with scheduled operation', function () {
           beforeEach(async function () {
-            ({ logs: this.logs } = await this.tlctrl.scheduleBatch(
+            ({ receipt: this.receipt, logs: this.logs } = await this.tlctrl.scheduleBatch(
               this.operation.targets,
               this.operation.values,
               this.operation.datas,
@@ -453,7 +452,7 @@ describe('TimelockController', function () {
                   this.operation.salt,
                   { from: executer }
                 ),
-                'Timelock: operation is not ready'
+                'TimelockController: operation is not ready'
               );
             });
           });
@@ -474,7 +473,7 @@ describe('TimelockController', function () {
                   this.operation.salt,
                   { from: executer }
                 ),
-                'Timelock: operation is not ready'
+                'TimelockController: operation is not ready'
               );
             });
           });
@@ -494,7 +493,6 @@ describe('TimelockController', function () {
                 this.operation.salt,
                 { from: executer }
               );
-              expectEvent(receipt, 'Executed', { id: this.operation.id });
               for (const i in this.operation.targets) {
                 expectEvent(receipt, 'CallExecuted', {
                   id: this.operation.id,
@@ -619,7 +617,7 @@ describe('TimelockController', function () {
           ZERO_BYTES32,
           web3.utils.randomHex(32),
         );
-        ({ logs: this.logs } = await this.tlctrl.schedule(
+        ({ receipt: this.receipt, logs: this.logs } = await this.tlctrl.schedule(
           this.operation.target,
           this.operation.value,
           this.operation.data,
@@ -648,21 +646,21 @@ describe('TimelockController', function () {
     it('prevent unauthorized maintenance', async function () {
       await expectRevert(
         this.tlctrl.updateDelay(0, { from: other }),
-        'TimelockController: sender requiers permission'
+        'TimelockController: restricted maintenance access'
       );
     });
 
     it('timelock scheduled maintenance', async function () {
-      const randomBN = web3.utils.randomHex(16);
+      const randomBN = web3.utils.toBN(web3.utils.randomHex(8)).toString();
       const operation = genOperation(
         this.tlctrl.address,
         0,
-        this.tlctrl.contract.methods.updateDelay(randomBN).encodeABI(),
+        this.tlctrl.contract.methods.updateDelay(randomBN.toString()).encodeABI(),
         ZERO_BYTES32,
         web3.utils.randomHex(32),
       );
 
-      await this.tlctrl.schedule(
+      const receipt1 = await this.tlctrl.schedule(
         operation.target,
         operation.value,
         operation.data,
@@ -671,8 +669,16 @@ describe('TimelockController', function () {
         MINDELAY,
         { from: proposer }
       );
+      expectEvent(receipt1, 'CallScheduled', {
+        id: operation.id,
+        index: web3.utils.toBN(0),
+        target: operation.target,
+        value: web3.utils.toBN(operation.value),
+        data: operation.data,
+        predecessor: operation.predecessor,
+      });
       await time.increase(MINDELAY);
-      const receipt = await this.tlctrl.execute(
+      const receipt2 = await this.tlctrl.execute(
         operation.target,
         operation.value,
         operation.data,
@@ -680,17 +686,96 @@ describe('TimelockController', function () {
         operation.salt,
         { from: executer }
       );
-      expectEvent(receipt, 'Executed', { id: operation.id });
-      expectEvent(receipt, 'CallExecuted', { id: operation.id });
-      expectEvent(receipt, 'MinDelayChange', { newDuration: web3.utils.toBN(randomBN), oldDuration: MINDELAY });
+      expectEvent(receipt2, 'CallExecuted', {
+        id: operation.id,
+        index: web3.utils.toBN(0),
+        target: operation.target,
+        value: web3.utils.toBN(operation.value),
+        data: operation.data,
+      });
+      expectEvent(receipt2, 'MinDelayChange', { newDuration: randomBN, oldDuration: MINDELAY });
 
-      expect(await this.tlctrl.viewMinDelay()).to.be.bignumber.equal(web3.utils.toBN(randomBN));
+      expect(await this.tlctrl.viewMinDelay()).to.be.bignumber.equal(randomBN);
+
+      console.log('Gas cost schedule:', receipt1.receipt.gasUsed);
+      console.log('Gas cost execute:', receipt2.receipt.gasUsed);
+    });
+  });
+
+  describe('dependency', function () {
+    beforeEach(async function () {
+      this.operation1 = genOperation(
+        ZERO_ADDRESS,
+        0,
+        web3.utils.randomHex(4),
+        ZERO_BYTES32,
+        web3.utils.randomHex(32),
+      );
+      this.operation2 = genOperation(
+        ZERO_ADDRESS,
+        0,
+        web3.utils.randomHex(4),
+        this.operation1.id,
+        web3.utils.randomHex(32),
+      );
+      await this.tlctrl.schedule(
+        this.operation1.target,
+        this.operation1.value,
+        this.operation1.data,
+        this.operation1.predecessor,
+        this.operation1.salt,
+        MINDELAY,
+        { from: proposer }
+      );
+      await this.tlctrl.schedule(
+        this.operation2.target,
+        this.operation2.value,
+        this.operation2.data,
+        this.operation2.predecessor,
+        this.operation2.salt,
+        MINDELAY,
+        { from: proposer }
+      );
+      await time.increase(MINDELAY);
+    });
+
+    it('cannot execute before dependency', async function () {
+      await expectRevert(
+        this.tlctrl.execute(
+          this.operation2.target,
+          this.operation2.value,
+          this.operation2.data,
+          this.operation2.predecessor,
+          this.operation2.salt,
+          { from: executer }
+        ),
+        'TimelockController: missing dependency'
+      );
+    });
+
+    it('can execute after dependency', async function () {
+      await this.tlctrl.execute(
+        this.operation1.target,
+        this.operation1.value,
+        this.operation1.data,
+        this.operation1.predecessor,
+        this.operation1.salt,
+        { from: executer }
+      );
+      await this.tlctrl.execute(
+        this.operation2.target,
+        this.operation2.value,
+        this.operation2.data,
+        this.operation2.predecessor,
+        this.operation2.salt,
+        { from: executer }
+      );
     });
   });
 
   describe('scenari', function () {
     it('call', async function () {
-      const randomBN = web3.utils.randomHex(16);
+      const randomBN = web3.utils.toBN(web3.utils.randomHex(8)).toString();
       const operation = genOperation(
         this.implementation2.address,
         0,
@@ -717,10 +802,15 @@ describe('TimelockController', function () {
         operation.salt,
         { from: executer }
       );
-      expectEvent(receipt, 'Executed', { id: operation.id });
-      expectEvent(receipt, 'CallExecuted', { id: operation.id });
+      expectEvent(receipt, 'CallExecuted', {
+        id: operation.id,
+        index: web3.utils.toBN(0),
+        target: operation.target,
+        value: web3.utils.toBN(operation.value),
+        data: operation.data,
+      });
 
-      expect(await this.implementation2.getValue()).to.be.bignumber.equal(web3.utils.toBN(randomBN));
+      expect(await this.implementation2.getValue()).to.be.bignumber.equal(randomBN);
     });
 
     it('call reverting', async function () {
@@ -850,8 +940,13 @@ describe('TimelockController', function () {
         operation.salt,
         { from: executer, value: 1 }
       );
-      expectEvent(receipt, 'Executed', { id: operation.id });
-      expectEvent(receipt, 'CallExecuted', { id: operation.id });
+      expectEvent(receipt, 'CallExecuted', {
+        id: operation.id,
+        index: web3.utils.toBN(0),
+        target: operation.target,
+        value: web3.utils.toBN(operation.value),
+        data: operation.data,
+      });
 
       expect(await web3.eth.getBalance(this.tlctrl.address)).to.be.bignumber.equal(web3.utils.toBN(0));
       expect(await web3.eth.getBalance(this.callreceivermock.address)).to.be.bignumber.equal(web3.utils.toBN(1));
